Toggling codegen packages for publishing
