+++ packages/sdk-codegen-utils/README.md
@@ -2,6 +2,9 @@
 
 This package contains utility routines for the SDK code generator projects
 
-The code in this package is really trivial and probably not pulling in as a dependency to anything.
+The code in this package is really trivial and should not be used directly in your own projects.
+
+**DEPRECATED AND NOT SUPPORTED**: Looker does not support direct use of this package, other than via direct use in the
+[SDK Codegen project](https://github.com/looker-open-source/sdk-codegen) repository.
 
 See the [SDK Codegen project](https://github.com/looker-open-source/sdk-codegen) repository for more information.
